refactor(package): packageUtil の計算部分を shared へ移設し特性化テストを追加
- states / parsePackageType / detectPackageTypes / getManuallyInstalledFiles /
  getInstalledVersionOfPackage / computePackagesStatus を
  src/shared/packageUtil.ts へ移設(electron 非依存の純ロジック)
- computePackagesStatus は旧 getPackagesStatus の計算部分
  (依存 |・競合 &・バージョン演算子・aviutl/exedit 疑似 ID の解決)。
  renderer 側は apm.json からのバージョン読み出しだけ残して委譲
- 特性化テスト 22 件(tmpdir でのインストール状態判定 7 件を含む)。
  isInstallable がバージョン指定付き ID を解決しない現行挙動も
  テストコメントで固定
